Unify perplexity calculations.
